export parsed composite specs
