+++ alpha/template/composite/composite.go
@@ -5,7 +5,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"io"
-	"net/http"
 	"net/url"
 	"os"
 	"path/filepath"
@@ -13,23 +12,6 @@ import (
 	"github.com/operator-framework/operator-registry/pkg/image"
 	"k8s.io/apimachinery/pkg/util/yaml"
 )
-
-type BuilderMap map[string]Builder
-
-type CatalogBuilderMap map[string]BuilderMap
-
-type builderFunc func(BuilderConfig) Builder
-
-type Template struct {
-	catalogFile        io.Reader
-	contributionFile   io.Reader
-	validate           bool
-	outputType         string
-	registry           image.Registry
-	registeredBuilders map[string]builderFunc
-}
-
-type TemplateOption func(t *Template)
 
 func WithCatalogFile(catalogFile io.Reader) TemplateOption {
 	return func(t *Template) {
@@ -79,10 +61,6 @@ func NewTemplate(opts ...TemplateOption) *Template {
 	return temp
 }
 
-type HttpGetter interface {
-	Get(url string) (*http.Response, error)
-}
-
 // FetchCatalogConfig will fetch the catalog configuration file from the given path.
 // The path can be a local file path OR a URL that returns the raw contents of the catalog
 // configuration file.
@@ -100,7 +78,7 @@ func FetchCatalogConfig(path string, httpGetter HttpGetter) (io.ReadCloser, erro
 		}
 	} else {
 		// Evalute remote catalog config
-		// If URi is valid, execute fetch
+		// If URI is valid, execute fetch
 		tempResp, err := httpGetter.Get(catalogURI.String())
 		if err != nil {
 			return nil, fmt.Errorf("fetching remote catalog config file %q: %v", path, err)
@@ -111,26 +89,37 @@ func FetchCatalogConfig(path string, httpGetter HttpGetter) (io.ReadCloser, erro
 	return tempCatalog, nil
 }
 
-// TODO(everettraven): do we need the context here? If so, how should it be used?
+func (t *Template) Parse() (*Specs, error) {
+	var s Specs
+
+	catalogSpec, err := t.parseCatalogsSpec()
+	if err != nil {
+		return nil, err
+	}
+	s.CatalogSpec = catalogSpec
+
+	contributionSpec, err := t.parseContributionSpec()
+	if err != nil {
+		return nil, err
+	}
+	s.ContributionSpec = contributionSpec
+
+	return &s, nil
+}
+
 func (t *Template) Render(ctx context.Context, validate bool) error {
-
-	catalogFile, err := t.parseCatalogsSpec()
+	specs, err := t.Parse()
 	if err != nil {
 		return err
 	}
 
-	contributionFile, err := t.parseContributionSpec()
-	if err != nil {
-		return err
-	}
-
-	catalogBuilderMap, err := t.newCatalogBuilderMap(catalogFile.Catalogs, t.outputType)
+	catalogBuilderMap, err := t.newCatalogBuilderMap(specs.CatalogSpec.Catalogs, t.outputType)
 	if err != nil {
 		return err
 	}
 
 	// TODO(everettraven): should we return aggregated errors?
-	for _, component := range contributionFile.Components {
+	for _, component := range specs.ContributionSpec.Components {
 		if builderMap, ok := (*catalogBuilderMap)[component.Name]; ok {
 			if builder, ok := builderMap[component.Strategy.Template.Schema]; ok {
 				// run the builder corresponding to the schema
