+++ alpha/template/composite/types.go
@@ -1,6 +1,12 @@
 package composite
 
-import "encoding/json"
+import (
+	"encoding/json"
+	"io"
+	"net/http"
+
+	"github.com/operator-framework/operator-registry/pkg/image"
+)
 
 type TemplateDefinition struct {
 	Schema string
@@ -26,4 +32,30 @@ type CustomConfig struct {
 	Command string
 	Args    []string
 	Output  string
+}
+
+type BuilderMap map[string]Builder
+
+type CatalogBuilderMap map[string]BuilderMap
+
+type builderFunc func(BuilderConfig) Builder
+
+type Template struct {
+	catalogFile        io.Reader
+	contributionFile   io.Reader
+	validate           bool
+	outputType         string
+	registry           image.Registry
+	registeredBuilders map[string]builderFunc
+}
+
+type TemplateOption func(t *Template)
+
+type Specs struct {
+	CatalogSpec      *CatalogConfig
+	ContributionSpec *CompositeConfig
+}
+
+type HttpGetter interface {
+	Get(url string) (*http.Response, error)
 }
